pyproject.toml
[project]
name = "library-catalog"
version = "1.0.0"
description = "Library Catalog REST API with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",

    # Database
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.3,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "greenlet (>=3.3.0,<4.0.0)",

    # Validation
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",

    # HTTP client
    "httpx (>=0.28.1,<0.29.0)",

    # Environment
    "python-dotenv (>=1.2.1,<2.0.0)",

    # Caching - Redis support
    "aiocache (>=0.12.3,<0.13.0)",
    "redis[hiredis] (>=5.0.0,<6.0.0)",

    # Authentication & Security
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.9,<0.1.0)",
    "bcrypt (>=4.0.1,<5.0.0)",  # ← ИСПРАВЛЕНО: 4.0.1 вместо 5.0.0
    "domain (>=0.1.4,<0.2.0)",
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.23.0,<0.24.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.8.5,<0.9.0)",
    "mypy (>=1.8.0,<2.0.0)",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
where = ["src"]